backend_py/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
httpx==0.27.2
pandas==2.2.3
numpy==2.1.1
websockets==13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.15.1
fastapi-users[sqlalchemy]==12.1.3

